flask_api/requirements.txt
flask
Flask-SQLAlchemy
Flask-Testing
SQLAlchemy
python-dotenv
requests
nose
